chore: i32 to u16

+++ src/main.rs
@@ -4,7 +4,7 @@ pub mod net;
 
 fn main() {
   println!("redis start");
-  let port: i32 = 6379;
+  let port: u16 = 6379;
   let mut server = Server::new("127.0.0.1", &port);
   server.start();
 }
